+++ Changelog.md
@@ -1,3 +1,6 @@
+- (***ADDB***) Fix reroll perception causing a NullReferenceException
+- 
+
 ### ToyBox DarkHeresy - Ver 0.1.9 (built for 0.4.0.749)
 - (***ADDB***) Fix saving sometimes not working (I'm not sure if this is caused by ToyBox or just a base game issue)
 - (***ADDB***) Fix Visual Size Override Slider defaulting to 0